test(admin): reorganize organization controller tests by action
Following the project's convention of placing integration tests under
test/controllers/admin/<controller>/<action>_test.rb.
